// src/common/enums/role.enum.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  MANAGER = 'MANAGER',
  STAFF = 'STAFF',
  WAREHOUSE_MANAGER = 'WAREHOUSE_MANAGER',
  PURCHASING_STAFF = 'PURCHASING_STAFF',
  FINANCE_ADMIN = 'FINANCE_ADMIN',
  FINANCE_MANAGER = 'FINANCE_MANAGER',
  FINANCE_STAFF = 'FINANCE_STAFF',
  PRODUCTION_MANAGER = 'PRODUCTION_MANAGER',
  PRODUCTION_SUPERVISOR = 'PRODUCTION_SUPERVISOR',
  PRODUCTION_STAFF = 'PRODUCTION_STAFF',
  SALES_STAFF = 'SALES_STAFF',
  AUDITOR = 'AUDITOR',
}